requirements.txt
python-telegram-bot==20.7
Flask
gspread
oauth2client
gunicorn
python-dotenv
